+++ test/web_watchdog/state_test.clj
@@ -102,10 +102,10 @@
     (testing "site > :email-notification > missing :format, throws"
       (throws-for-updated-config #(dissoc-nested-key % [:sites 0 :email-notification :format])))
     (testing "site > :email-notification > valid :format values, passes"
-      (doseq [val ["old-new" "inline-diff"]]
+      (doseq [val ["old-new" "new-only" "inline-diff"]]
         (passes-for-updated-config #(assoc-in % [:sites 0 :email-notification :format] val))))
     (testing "site > :email-notification > invalid :format values, throws"
-      (doseq [val [:old-new :inline-diff nil "bogus"]]
+      (doseq [val [:old-new :new-only :inline-diff nil "bogus"]]
         (throws-for-updated-config #(assoc-in % [:sites 0 :email-notification :format] val)))
       (throws-for-updated-config #(dissoc-nested-key % [:sites 0 :email-notification :format])))
     (testing "site > :email-notification > valid :condition, passes"